refactor main event component card
